pyproject.toml
[project]
name = "bncapi"
version = "0.1.0"
description = "Backend API for bnc game"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-ninja>=1.4.3",
    "django-cors-headers>=4.7.0",
    "cryptography>=45.0.6",
    "django-activity-stream>=2.0.0",
    "djangorestframework>=3.16.0",
    "pydantic[email]>=2.11.7",
    "django-silk>=5.4.0",
    "channels[daphne]>=4.3.1",
    "channels-redis>=4.3.0",
    "psycopg2-binary>=2.9.10",
    "httpx>=0.28.1",
    "gunicorn>=23.0.0",
    "dj-database-url>=3.0.1",
    "uvicorn>=0.35.0",
    "whitenoise[brotli]>=6.9.0",
    "jsonpickle>=4.1.1",
    "schedule>=1.2.2",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
    "setuptools>=80.9.0",
    "wheel>=0.45.1",
    "build>=1.3.0",
    "twine>=6.1.0",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "scripts",
]

line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "N", "LOG", "G", "FBT", "B", "ICN"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
